api/pyproject.toml
[project]
name = "api"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "core",
    "fastapi>=0.120.2",
    "loguru>=0.7.3",
    "pydantic-settings>=2.11.0",
    "uvicorn>=0.25.0",
]

[tool.pyright]
exclude = ["**/node_modules", ".venv"]
include = ["./src/"]
venvPath = "."
venv = ".venv"

[tool.black]
line-length = 88
target-version = ["py312"]

[tool.uv.sources]
core = { path = "../core" }
